3 mayo 22.30

+++ 3raEval/WorkspaceSTS/PilarAguilarTiendaOnline01/src/main/java/com/salesianostriana/dam/pilaraguilartiendaonline01/model/BasicUser.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class BasicUser {
 	
 	@Id @GeneratedValue
-	private int basicUserId;
+	private Long basicUserId;
 	
 	private String basicUserName;
 	private String basicUserPassword;
