+++ 3raEval/WorkspaceSTS/PilarAguilarTiendaOnline01/src/main/java/com/salesianostriana/dam/pilaraguilartiendaonline01/model/Order.java
@@ -26,7 +26,8 @@ public class Order {
 
 	@Id
 	@GeneratedValue
-	private int orderId;
+	private Long orderId;
+	
 	
 	private Date orderDate;
